// src/lib/enums.ts
export enum CurrencyType {
	USD = "USD",
	GBP = "GBP",
	BDT = "BDT",
}

export enum EmployeeStatus {
	ACTIVE = "active",
	INACTIVE = "inactive",
	TERMINATED = "terminated",
	ON_LEAVE = "on_leave",
	PROBATION = "probation",
}

export enum JobOpeningStatus {
	OPEN = "open",
	CLOSED = "closed",
}

export enum LeaveRequestStatus {
	PENDING = "pending",
	APPROVED = "approved",
	REJECTED = "rejected",
}

export enum LeaveType {
	VACATION = "vacation",
	SICK_LEAVE = "sick_leave",
	PARENTAL_LEAVE = "parental_leave",
	MEDICAL_LEAVE = "medical_leave",
}